feat: Move code in __init__ into separate modules
This allow the module to have a consistent code separation.
Additionally, it will help in the future in case of changes to one part
of the application, e.g. version parsing, without affecting others.
